modify plugin-tray  storage show model
