+++ plugin-tray/ukuitray.h
@@ -44,34 +44,18 @@
 #include "../panel/iukuipanel.h"
 #include "../panel/customstyle.h"
 #include "../panel/ukuicontrolstyle.h"
-#include "traystorage.h"
+#include "ukuitraystrage.h"
 
 class TrayIcon;
 class QSize;
 namespace UKUi {
 class GridLayout;
 }
-class UKUiStorageWidget;
+
 /**
  * @brief This makes our trayplugin
  */
 class UKUITrayPlugin;
-enum storageBarStatus{ST_HIDE,ST_SHOW};
-/**
- * @brief This makes our storage
- */
-class UKUIStorageFrame:public QWidget
-{
-    Q_OBJECT
-public:
-    UKUIStorageFrame(QWidget* parent =0);
-    ~UKUIStorageFrame();
-protected:
-    bool eventFilter(QObject *, QEvent *);
-    void paintEvent(QPaintEvent *event)override;
-private:
-    Atom _NET_SYSTEM_TRAY_OPCODE;
-};
 
 class UKUITray: public QFrame, QAbstractNativeEventFilter
 {
@@ -83,7 +67,6 @@ public:
 
     QSize iconSize() const { return mIconSize; }
     void setIconSize();
-    void setStorageBar(TrayStorage *pTys);
     bool nativeEventFilter(const QByteArray &eventType, void *message, long *);
     UKUITrayPlugin *mPlugin;
 
@@ -160,12 +143,4 @@ private:
     QGSettings *settings;
 };
 
-class UKUiStorageWidget:public QWidget
-{
-public:
-    UKUiStorageWidget();
-    ~UKUiStorageWidget();
-protected:
-    void paintEvent(QPaintEvent*);
-};
 #endif
